requirements.txt
pandas
sqlalchemy
psycopg2-binary
google-cloud-storage
pyarrow
python-dotenv
fastapi
uvicorn
pydantic